pyproject.toml
[tool.poetry]
name = "bld-ai-tasks"
version = "0.1.0"
description = ""
authors = ["AhmedOsama198 <ahmed.osama1982002@gmail.com>"]
readme = "README.md"
packages = [{include = "bld"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
django-model-utils = "^4.2.0"
pillow = "^9.3.0"
django-imagekit = "^4.1.0"
knox = "^0.1.14"
cryptography = "^38.0.3"
boto3 = "^1.26.7"
django-rest-knox = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
